Move Sfdc::Api under Embulk::Input
